requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.42.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-datatables-view==1.19.1
#django-sendgrid-v5==0.9.0
django-tempus-dominus==5.1.2.13
flake8==3.8.3
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
isort==5.5.3
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
#mysqlclient==2.0.1
oauthlib==3.1.0
#psycopg2==2.8.6
#psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-decouple==3.3
python-http-client==3.3.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
sendgrid==6.6.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
uuid==1.30
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.1.0
